+++ .gitignore
@@ -26,3 +26,4 @@ dist-ssr
 *.sln
 *.sw?
 server/config/oracle.js
+
