+++ package.json
@@ -34,7 +34,7 @@
     "typeorm": "^0.3.17"
   },
   "devDependencies": {
-    "@types/node": "^20.9.5",
+    "@types/axios": "^0.14.0",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
