+++ src/App.tsx
@@ -4,16 +4,16 @@ import './App.css';
 import { Route, Routes } from 'react-router-dom';
 
 import NavigationBar from './components/NavigationBar';
-import Home from './pages/Home';
-import Content from './pages/Content';
-import Style from './pages/Style';
-import Tpo from './pages/Tpo';
+import Home from './Pages/Home';
+import Content from './Pages/Content';
+import Style from './Pages/Style';
+import Tpo from './Pages/Tpo';
 // import Footer from '@components/footer';
-import Detail from './pages/Detail';
-import Sign from './pages/Sign';
+import Detail from './Pages/Detail';
+import Sign from './Pages/Sign';
 import SignUp from './components/SignUp';
-import ContentDetail from 'pages/ContentDetail';
-import MyPage from './pages/MyPage';
+import ContentDetail from 'Pages/ContentDetail';
+import MyPage from './Pages/MyPage';
 
 const App: React.FC = () => {
   const [userInput, setUserInput] = useState('');
